+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * rework Draw.io icon and add an alternative icon.
 * rework Dependabot icons (PNG icon to SVG).
 * support Stylus (browser extension) `*.styl` files.
+* support Mergify YML files.
 
 ## 1.55.0 (2021/05/29)
 * rework support of `*.spec.tsx` files.